+++ docs/changes.md
@@ -5,6 +5,9 @@
 - improvements in doc strings
 
 ## Breaking changes
+- improvements to `shapes.Ellipsoid()` and bug fixes in #978 by @daniel-a-diaz
+- improvements to `pointcloud.pca_ellipsoid()` and bug fixes
+- improvements to `pointcloud.pca_ellipse()` and bug fixes
 
 ### Renaming